v1/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.6
django-bootstrap5==25.1
idna==3.10
mysqlclient==2.2.7
pillow==11.1.0
pycparser==2.22
pyOpenSSL==25.0.0
python-decouple==3.8
RapidFuzz==3.12.2
requests==2.32.3
setuptools==78.1.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
wheel==0.45.1
